Removed unused file.
